Draft Metadata

+++ src/AudioCodecs/CodecMP3LAME.h
@@ -106,7 +106,7 @@ public:
 	
 	// convert PCM data to convert into MP3
 	size_t write(const void *in_ptr, size_t in_size){
-		LOGD("write %d bytes", in_size);
+		LOGD("write %d bytes", (int) in_size);
 		return enc->write((uint8_t*)in_ptr, in_size);
 	}
